Security

+++ View/EmployeeView.cs
@@ -4,18 +4,21 @@ using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
 using Bogus;
+using Lab1.Context;
 
 namespace Lab1.View
 {
     public partial class EmployeeView : Form
     {
         private readonly IEmployeeController _employeeController;
+        private readonly SecurityContext _securityContext;
         private Employee _currentEmployee;
 
         public EmployeeView(IEmployeeController employeeController)
         {
             InitializeComponent();
-            List<string> enumNames = new List<string>(Enum.GetNames(typeof(Position)));
+            _securityContext = SecurityContext.Instance;
+            var enumNames = new List<string>(Enum.GetNames(typeof(Position)));
             enumNames.Insert(0, "Not chosen");
             positionComboBox.DataSource = enumNames;
             _employeeController = employeeController;
@@ -24,22 +27,29 @@ namespace Lab1.View
 
         private void UpdateGrid()
         {
-            employeesGrid.DataSource = _employeeController.GetAll();
+            if (_securityContext.IsAccessible(_employeeController, nameof(_employeeController.GetAll)))
+            {
+                employeesGrid.DataSource = _employeeController.GetAll();
+            }
         }
 
         private void findButton_Click(object sender, EventArgs e)
         {
             UpdateCurrentEmployee();
-            List<Employee> foundEmployees = _employeeController.Find(_currentEmployee);
-            employeesGrid.DataSource = foundEmployees;
+            if (_securityContext.IsAccessible(_employeeController, nameof(_employeeController.Find)))
+            {
+                var foundEmployees = _employeeController.Find(_currentEmployee);
+                employeesGrid.DataSource = foundEmployees;
+            }
         }
+
         private Employee ParseFields()
         {
-            string id = idTextBox.Text;
-            string firstName = firstNameTextBox.Text;
-            string secondName = secondNameTextBox.Text;
-            string position = positionComboBox.Text;
-            string address = addressTextBox.Text;
+            var id = idTextBox.Text;
+            var firstName = firstNameTextBox.Text;
+            var secondName = secondNameTextBox.Text;
+            var position = positionComboBox.Text;
+            var address = addressTextBox.Text;
             return new Employee(id, firstName, secondName,
                 position, address);
         }
@@ -52,22 +62,33 @@ namespace Lab1.View
         private void addButton_Click(object sender, EventArgs e)
         {
             UpdateCurrentEmployee();
-            _employeeController.Add(_currentEmployee);
+            if (_securityContext.IsAccessible(_employeeController, nameof(_employeeController.Add)))
+            {
+                _employeeController.Add(_currentEmployee);
+            }
+
             UpdateGrid();
         }
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
             UpdateCurrentEmployee();
-            _employeeController.Delete(_currentEmployee);
+            if (_securityContext.IsAccessible(_employeeController, nameof(_employeeController.Delete)))
+            {
+                _employeeController.Delete(_currentEmployee);
+            }
+
             UpdateGrid();
         }
 
         private void editButton_Click(object sender, EventArgs e)
         {
-            _employeeController.Edit(_currentEmployee, ParseFields());
-            UpdateCurrentEmployee();
-            UpdateGrid();
+            if (_securityContext.IsAccessible(_employeeController, nameof(_employeeController.Edit)))
+            {
+                _employeeController.Edit(_currentEmployee, ParseFields());
+                UpdateCurrentEmployee();
+                UpdateGrid();
+            }
         }
 
         private void employeesGrid_CellClick(object sender, DataGridViewCellEventArgs e)
@@ -76,7 +97,8 @@ namespace Lab1.View
             {
                 return;
             }
-            DataGridView dataGrid = (DataGridView)sender;
+
+            var dataGrid = (DataGridView) sender;
             var row = dataGrid.Rows[e.RowIndex];
 
             idTextBox.Text = row.Cells[0].Value.ToString();
@@ -94,17 +116,22 @@ namespace Lab1.View
 
         private void fakeInfoButton_Click(object sender, EventArgs e)
         {
-            Employee employee = new Employee();
+            if (!_securityContext.IsAccessible(_employeeController, nameof(_employeeController.Add)))
+            {
+                return;
+            }
+
+            var employee = new Employee();
             for (int i = 1; i <= 100; i++)
             {
-                Faker faker = new Faker();
-                Person person = faker.Person;
+                var faker = new Faker();
+                var person = faker.Person;
                 employee.Id = i;
                 employee.FirstName = person.FirstName;
                 employee.SecondName = person.LastName;
-                employee.Position = (Position)(i % 3);
+                employee.Position = (Position) (i % 3);
                 var address = person.Address;
-                employee.Address = string.Join(", ", address.Street, 
+                employee.Address = string.Join(", ", address.Street,
                     address.City, address.ZipCode, address.State);
                 _employeeController.Add(employee);
                 UpdateGrid();
